cast to appropriate data type
git-svn-id: http://repository.devel.hbm.com/svn/quantumx/trunk@22515 7cbc33cc-6077-418a-a044-ec03dd7e8bdb

+++ communication/multicastserver.cpp
@@ -185,7 +185,7 @@ namespace hbm {
 
 		ssize_t MulticastServer::receiveTelegram(void* msgbuf, size_t len, int& adapterIndex, std::chrono::milliseconds timeout)
 		{
-			int milliSeconds = timeout.count();
+			int milliSeconds = static_cast < int > (timeout.count());
 
 
 			int retval;
